stand up 'Hello, World" demo rename package
