+++ .gitignore
@@ -26,3 +26,4 @@ local.properties
 .classpath
 .settings/
 .loadpath
+/target/
