+++ src/main/java/com/angmas/mutation/restservice/DemoApplication.java
@@ -1,4 +1,4 @@
-package com.springFileMutation.demo;
+package com.angmas.mutation.restservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
