+++ src/test/java/com/angmas/mutation/restservice/GreetingTest.java
@@ -1,0 +1,24 @@
+package com.angmas.mutation.restservice;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+class GreetingTest {
+
+	@BeforeEach
+	void setUp() throws Exception {
+	}
+
+	@AfterEach
+	void tearDown() throws Exception {
+	}
+
+	@Test
+	void returnsHelloWorld() {
+//		assertEquals("");
+	}
+
+}
